% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Align Spatial Data}
\usage{
align(
  spatial_data,
  base_layer,
  resample_method = "bilinear",
  show_plots = FALSE,
  verbose = TRUE,
  output_file = NULL
)
}
\arguments{
\item{spatial_data}{A spatial object to be aligned.}

\item{base_layer}{A spatial object to use as a reference for extent, cell size, grid offset, and final CRS.}

\item{resample_method}{Character. Method used for estimating the new cell values. One of "near", "bilinear", "cubic", "cubicspline", "lanczos", "sum", "min", "q1", "med", "q3", "max", "average", "mode", "rms". Defaults to "bilinear".}

\item{show_plots}{Logical, whether to create and display before and after plots. Defaults to FALSE.}

\item{verbose}{Logical, whether to provide detailed logging. Defaults to TRUE.}

\item{output_file}{Optional. A file path to save the aligned spatial data.}
}
\value{
A spatial object aligned with the base layer's CRS, extent, cell size, and grid offset.
}
\description{
\code{align} prepares a spatial object by aligning its CRS, extent, cell size, and grid offset based on a provided base layer.
}
\examples{
aligned_spatial_data <- align(spatial_data = spatial_data1, base_layer = spatial_data2, resample_method = "bilinear", show_plots = TRUE)
}
