% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_properties.R
\name{verify_properties}
\alias{verify_properties}
\title{Verify Spatial Data Properties}
\usage{
verify_properties(spatial_data1, spatial_data2)
}
\arguments{
\item{spatial_data1}{A spatial object to be compared.}

\item{spatial_data2}{A spatial object to be compared against.}
}
\description{
\code{verify_properties} checks the properties of two spatial objects, such as CRS, extent, resolution, and data type, and provides a comparison summary of these properties.
If the spatial objects are rasters, it also provides a separate table for resolution.
}
\examples{
verify_properties(spatial_data1, spatial_data2)
}
